Python2 to Python3 conversion fixes
lnst/Common/Daemon.py
===================================
 Issue#191 - daemonize doesn't work with py3
- Replaces file method with open method instead

lnst/Controller/SlavePool.py
===================================
 Unable to match specific test cases
- Deleted iter from sorted list, as iterator object not needed

lnst/Controller/Task.py
===================================
 Hash creation was unsuccessful due to type mismatch
- hashlib.sha1 is expecting str, instead of unicode object
- Changed encoding to utf-8 str

lnst/Slave/NetTestSlave.py
===================================
 Top-level exception was thrown instead of specific one, could catch
 exception, that wasn't meant for Slave process
- Created specific SystemCallException
- Throwing and expecting SystemCallException instead of general one

recipes/regression_tests/phase3/ipsec_esp_aead.py
===================================
 Type mismatch
- py3 divison is float by default, added explicit conversion to int

test_modules/Netperf.py
===================================
 Unable to interrupt netperf server instance
- Created new InterruptException
- Created new method wait_for_interrupt to catch sigint, instead
of waiting for exception, that's no longer thrown with PEP 475
- py3 divison is float by default, added explicit conversion to int

+++ recipes/regression_tests/phase3/ipsec_esp_aead.py
@@ -14,7 +14,7 @@ from lnst.RecipeCommon.PerfRepo import generate_perfrepo_comment
 #lenth param is in bits
 def generate_key(length):
     key = "0x"
-    key = key + (length/8) * "0b"
+    key = key + (int(length/8)) * "0b"
     return key
 
 algorithm = []
